feat: always bundle V2 functions
