+++ tests/fixtures-esm/v2-api-esm-mixed-modules/function-ts.ts
@@ -1,9 +1,9 @@
 import cjs from 'cjs-module'
 import esm from 'esm-module'
 
-import { name as helper1 } from './lib/helper1.ts'
+import { name as helper1 } from './lib/helper1.js'
 import { name as helper2 } from './lib/helper2.js'
-import { name as helper3 } from './lib/helper3'
+import { name as helper3 } from './lib/helper3.js'
 import { name as helper4 } from './lib/helper4.js'
 import { name as helper5 } from './lib/helper5.mjs'
 import { name as helper6 } from './lib/helper6.js'
